[Refactor]: 레시피 도메인 모델 일부 프로퍼티 데이터 타입 변경

+++ RecipeForT/Entity/Recipe.swift
@@ -18,11 +18,11 @@ struct Recipe: Identifiable {
     /// 대표 이미지 URL
     let imageURL: URL?
     /// n인분, 인당 식사 분량
-    let servingsCount: UInt8
+    let servingsCount: Decimal
     /// 조리비용
     let cost: Decimal
     /// 조리시간(분)
-    let cookingTime: UInt8
+    let cookingTime: Decimal
     /// 레시피에 대한 설명
     let description: String
     /// 재료 목록
@@ -33,9 +33,9 @@ struct Recipe: Identifiable {
         authorID: UInt64 = .zero,
         name: String = "",
         imageURL: URL? = nil,
-        servingsCount: UInt8 = .zero,
+        servingsCount: Decimal = .zero,
         cost: Decimal = .zero,
-        cookingTime: UInt8 = .zero,
+        cookingTime: Decimal = .zero,
         description: String = "",
         ingredients: [Ingredient] = []
     ) {
